wuying-agentbay-sdk-main/python/pyproject.toml
[tool.poetry]
name = "wuying-agentbay-sdk"
version = "0.9.2"
description = "Python SDK for interacting with the Wuying AgentBay cloud runtime environment"
readme = ["README.md", "CHANGELOG.md"]
packages = [{include = "agentbay"}]
authors = ["Wuying AI Team <wuying-ai-team@alibabacloud.com>"]
homepage = "https://github.com/aliyun/wuying-agentbay-sdk"
repository = "https://github.com/aliyun/wuying-agentbay-sdk"
documentation = "https://github.com/aliyun/wuying-agentbay-sdk/docs"
license = "Apache-2.0"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

[tool.poetry.dependencies]
python = ">=3.10"
darabonba-core = ">=1.0.0,<2.0.0"
alibabacloud-tea-openapi = ">= 0.4.0rc3"
cryptography = ">=44.0.0"
httpx = "^0.28.1"
alibabacloud-tea-util = "^0.3.13"
python-dotenv = "^1.0.0"
pydantic = "^2.0"
playwright = ">=1.5.0"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.10.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
